Abstraction for different metadata collection strategies.
