+++ lib/galaxy/tools/actions/metadata.py
@@ -2,8 +2,8 @@ import logging
 import os
 from json import dumps
 
-from galaxy.datatypes.metadata import JobExternalOutputMetadataWrapper
 from galaxy.jobs.datasets import DatasetPath
+from galaxy.metadata import get_metadata_compute_strategy
 from galaxy.util.odict import odict
 from . import ToolAction
 
@@ -74,7 +74,7 @@ class SetMetadataToolAction(ToolAction):
         job_working_dir = app.object_store.get_filename(job, base_dir='job_work', dir_only=True, extra_dir=str(job.id))
         datatypes_config = os.path.join(job_working_dir, 'registry.xml')
         app.datatypes_registry.to_xml_file(path=datatypes_config)
-        external_metadata_wrapper = JobExternalOutputMetadataWrapper(job)
+        external_metadata_wrapper = get_metadata_compute_strategy(app, job.id)
         cmd_line = external_metadata_wrapper.setup_external_metadata(dataset,
                                                                      sa_session,
                                                                      exec_dir=None,
